Fix index return to respond function to match new models
