+++ activities/views.py
@@ -11,6 +11,7 @@ from django.views import generic
 from django.middleware.csrf import get_token
 from activities.decorator import login_required
 from django.views.decorators.http import require_POST
+from django.db.models import F, ExpressionWrapper, IntegerField
 
 
 class IndexView(generic.ListView):
@@ -26,12 +27,21 @@ class IndexView(generic.ListView):
 
         Queryset is order by date that the activity took place.(earlier to later)
         """
-        return models.Activity.objects.filter(date__gte=timezone.now()).order_by("date")
+        query = models.Activity.objects.filter(date__gte=timezone.now()).order_by("date")
+                            
+        return query
 
+        
     def render_to_response(self, context: Dict[str, Any], **response_kwargs: Any) -> JsonResponse:
         """Send out JSON response to Vue for Activity Index."""
+        
+        get_people = lambda act_id: models.Activity.objects.get(pk=act_id).people 
+        
         activities = list(self.get_queryset().values(
-            "id", "name", "detail", "date", "max_people", "people"))
+                "id", "name", "detail", "date", "max_people"
+            ))
+        
+        activities = [act | {"people": get_people(act["id"])} for act in activities]
         return JsonResponse(activities, safe=False)
